docs + changes
