+++ docs/lang.md
@@ -1,0 +1,10 @@
+## FaceitKt DIY documentation
+
+
+### If you understand this, click [**here**](./eng/welcome.md)
+
+
+### Если вы понимаете это, [**кликните сюда**](./ru/welcome.md)
+
+
+~~Choose anything if you understand both, they are equal bruh~~
