--- src/main/kotlin/Main.kt
@@ -1,5 +1,0 @@
-fun main(){
-    val user: FaceitUser = FaceitUser("s1mple", FaceitUser.Game.csgo)
-    println(user.gameStats)
-
-}
